git-svn-id: http://192.168.20.11/sequoiadb/trunk@11771 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/pmd/pmdSignalHandler.cpp
@@ -111,7 +111,7 @@ namespace engine
       } ;
 
       if ( signum > 0 &&
-           signum < sizeof(s_signalHandleMap)/sizeof(pmdSignalInfo) )
+           (UINT32)signum < sizeof(s_signalHandleMap)/sizeof(pmdSignalInfo) )
       {
          return s_signalHandleMap[ signum ] ;
       }
